+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+## [0.1.5] - 2022-07-23
 ### Added
 
 - Upload attachments (currently image-only).
@@ -19,6 +21,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Reply message not being cleared when switching channels.
 - Replies not showing correctly.
 - Another attempt at fixing application freeze
+- Non-adaptive UI in the error dialog
 
 ### Development
 
@@ -43,5 +46,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Fixed cargo clippy.
 - Updated presage to official repository.
 
-[Unreleased]: https://gitlab.com/Schmiddiii/flare/-/compare/0.1.4...master
+[Unreleased]: https://gitlab.com/Schmiddiii/flare/-/compare/0.1.5...master
+[0.1.5]: https://gitlab.com/Schmiddiii/flare/-/compare/0.1.4...0.1.5
 [0.1.4]: https://gitlab.com/Schmiddiii/flare/-/compare/0.1.3...0.1.4